fixed pareva integration
